AIPL-748: [feature] add tel for user

+++ app/model/user.js
@@ -10,7 +10,7 @@ module.exports = app => {
     gender: {type: STRING(255), defaultValue: ''}, // 性别
     country: {type: STRING(255), defaultValue: ''}, // 国家
     city: {type: STRING(255), defaultValue: ''}, // 城市
-    tel: INTEGER,// 电话
+    tel: INTEGER(11),// 电话
     updated_at: DATE, // 更新时间
   }, {
     // 是否需要增加createdAt、updatedAt、deletedAt字段
