+++ database/migrations/20181214092547-migration.js
@@ -12,7 +12,7 @@ module.exports = {
       gender: STRING(255), // 性别
       country: STRING(255), // 国家
       city: STRING(255), // 城市
-      tel: INTEGER,
+      tel: INTEGER(11),
       updated_at: DATE, // 更新时间
       dTime: DATE, // 删除时间
     });
